moved date extension to core and added some extension methods
